@@ -1,16 +1,19 @@
 package shoveler
 
 import (
-	"github.com/spf13/viper"
-	"github.com/stretchr/testify/assert"
 	"path"
 	"strconv"
 	"testing"
 	"time"
+
+	"github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
+	"github.com/stretchr/testify/assert"
 )
 
 // TestQueueInsert tests the good validation
 func TestQueueInsert(t *testing.T) {
+	log = logrus.New()
 	queuePath := path.Join(t.TempDir(), "shoveler-queue")
 	viper.Set("queue_directory", queuePath)
 	queue := NewConfirmationQueue()
